misc/elfd/elf.c: enable dumping symbols from Ubuntu 16.04 executables
Back when this program was first written, sometime on or before
1996-12-14, Linux ELF (32) executables always had a SHT_HASH section
that pointed to the symbol table and string table.  That seems to have
changed sometime before the earl 2000s but this program was never
updated to deal with executables without a SHT_HASH section.
Now it will also try and get the symbol/string table from
the SHT_SYMTAB and SHT_STRTAB sections.